fix memory leak in Matrix#view, Matrix#row_view, Matrix#column_view, Vector#view
in gslng_extensions.cpp, the view methods were allocating a new matrix or
vector object and and overwriting its contents with the information from
the view.  but the allocation via gsl_vector_alloc() included not only the
header struct, but also the memory for a matrix or vector, and  the pointer
to that memory was then overwritten and the memory itself never freed.  so
instead of allocating the matrix or vector object via gsl_vector_alloc(),
allocating only the header struct via ordinary malloc().

memory leak fix for vector views same as for matrix views